Switch service store with instance store
I separated the service instance store and the service store to make
the code easier to follow. The other alternative would be to make new
getAll and get methods on the existing service store for instances vs
services and store both in separate places which would be more code.
This way keeps all stores in a similar state.
